Bookmark system fully implemented
No more work needs to be done on bookmarks. The views change or are disabled entirely, according to BookmarkPolicies

Created root user for complete access to everything on the database

All bookmark views have been updated to implement the new BookmarkPolicies to hide elements from specific users

BookmarkController now uses the BookmarksPolicy to determine whether the user has access to the bookmark or not.